don't "need" to enable dnssec and validation default is yes not auto
Thanks to Tony Finch for reminder (I had already noted this before
but the fix was misplaced):

dnssec-enable is turned on by default, so I recommend not putting it in
your config file. (Putting it in might tempt people too turn it off which
stops named implementing the protocol correctly.)

The default setting for dnssec-validation is yes NOT auto.
"dnssec-validation yes" does not actually turn on validation unless you
also configure a trust anchor, and there is none configured by default.