Maintenance updates to the Validation chapter
* Updated text to take account of the fact that DNSSEC validation is
now enabled by default and that even if validation is disabled, BIND
always requests DNSSEC-related information (i.e. the DO bit is always
set).
* Updated examples to use "ftp.isc.org" instead of "www.isc.org".
(Following changes to the ISC web site, the latter is now a CNAME to a
name in a zone zone that is not signed, something that would
significantly complicate the examples and explanations.)
* Update text to replace the "managed-keys" and "trusted-keys"
statements with the appropriate "trust-anchors" syntax.
* Updated examples to show output using 9.16.0 "dig" (which is
virtually unchanged from 9.10.0 used in previous versions of the guide
apart from the version number).  These changes also made signature
inception and expiration dates more recent.