Maintenance updates to the Troubleshooting chapter
Most of the updates concerned changing the examples, displaying what
would be seen with a later version of "dig" (9.16.0 is used here).
Apart from the version number, this output from version of "dig"
includes the DNS cookie.

At this stage, the main troubleshooting example has been commented
out.  This example - of which the problem is that a key configured on
the resolver is not a match for a KSK for a zone on an authoritative
server - is not now realistic. Typical sites only have one key (the
root key) configured and even that is now managed by default in BIND.
A later update to this chapter may add another example.